Fixes #15405 - Moved puppet UI to helpers
