+++ app/helpers/hosts_and_hostgroups_helper.rb
@@ -7,26 +7,6 @@ module HostsAndHostgroupsHelper
     name
   end
 
-  def parent_classes(obj)
-    return obj.hostgroup.classes if obj.is_a?(Host::Base) && obj.hostgroup
-    return obj.is_root? ? [] : obj.parent.classes if obj.is_a?(Hostgroup)
-    []
-  end
-
-  def accessible_puppet_ca_proxies(obj)
-    list = accessible_resource_records(:smart_proxy).with_features("Puppet CA").to_a
-    current = obj.puppet_ca_proxy
-    list |= [current] if current.present?
-    list
-  end
-
-  def accessible_puppet_proxies(obj)
-    list = accessible_resource_records(:smart_proxy).with_features("Puppet").to_a
-    current = obj.puppet_proxy
-    list |= [current] if current.present?
-    list
-  end
-
   def domain_subnets(type)
     accessible_related_resource(@domain, :subnets, :where => {:type => type})
   end
@@ -61,82 +41,5 @@ module HostsAndHostgroupsHelper
                 },
                 { :help_inline   => :indicator }
             ).html_safe
-  end
-
-  def host_puppet_environment_field(form, select_options = {}, html_options = {})
-    select_options = {
-      :include_blank => true,
-      :disable_button => _(INHERIT_TEXT),
-      :disable_button_enabled => inherited_by_default?(:environment_id, @host),
-      :user_set => user_set?(:environment_id)}.deep_merge(select_options)
-
-    html_options = {
-      :data => {
-        :url => hostgroup_or_environment_selected_hosts_path,
-        :host => {
-          :id => @host.id
-        }
-      }}.deep_merge(html_options)
-
-    puppet_environment_field(
-      form,
-      accessible_resource(@host, :environment),
-      select_options,
-      html_options)
-  end
-
-  def hostgroup_puppet_environment_field(form, select_options = {}, html_options = {})
-    select_options = {
-      :include_blank => blank_or_inherit_f(form, :environment)
-    }.deep_merge(select_options)
-
-    html_options = {
-      :data => {
-        :url => environment_selected_hostgroups_path
-      }}.deep_merge(html_options)
-
-    puppet_environment_field(
-      form,
-      accessible_resource(@hostgroup, :environment),
-      select_options,
-      html_options)
-  end
-
-  def puppet_environment_field(form, environments_choice, select_options = {}, html_options = {})
-    html_options = {
-      :onchange => 'update_puppetclasses(this)',
-      :help_inline => :indicator}.deep_merge(html_options)
-
-    select_f(
-      form,
-      :environment_id,
-      environments_choice,
-      :id,
-      :to_label,
-      select_options,
-      html_options)
-  end
-
-  def interesting_puppetclasses(obj)
-    classes = obj.all_puppetclasses
-    classes_ids = classes.reorder('').pluck('puppetclasses.id')
-    smart_vars = VariableLookupKey.reorder('').where(:puppetclass_id => classes_ids).uniq.pluck(:puppetclass_id)
-    class_vars = PuppetclassLookupKey.reorder('').joins(:environment_classes).where(:environment_classes => { :puppetclass_id => classes_ids }).uniq.pluck('environment_classes.puppetclass_id')
-    klasses    = (smart_vars + class_vars).uniq
-
-    classes.where(:id => klasses)
-  end
-
-  def puppetclasses_tab(puppetclasses_receiver)
-    return unless accessible_puppet_proxies(puppetclasses_receiver).present?
-    content_tag(:div, :class => 'tab-pane', :id => 'puppet_klasses') do
-      if @environment.present? ||
-          @hostgroup.present? && @hostgroup.environment.present?
-        render 'puppetclasses/class_selection', :obj => puppetclasses_receiver
-      else
-        alert(:class => 'alert-info', :header => _('Notice'),
-              :text => _('Please select an environment first'))
-      end
-    end
   end
 end
