+++ app/helpers/pagelets_helper.rb
@@ -21,7 +21,7 @@ module PageletsHelper
   def render_tab_content_for(mountpoint, opts = {})
     result = ""
     pagelets_for(mountpoint).each do |pagelet|
-      next unless pagelet.onlyif.call opts[:subject]
+      next unless pagelet.onlyif.call(opts[:subject], self)
       result += "<div id='#{pagelet.id}' class='tab-pane'>"
       result += render_pagelet(pagelet, opts)
       result += "</div>"
@@ -32,14 +32,14 @@ module PageletsHelper
   def render_tab_header_for(mountpoint, opts = {})
     result = ""
     pagelets_for(mountpoint).each do |pagelet|
-      next unless pagelet.onlyif.call opts[:subject]
+      next unless pagelet.onlyif.call(opts[:subject], self)
       result += "<li><a href='##{pagelet.id}' data-toggle='tab'>#{_(pagelet.name)}</a></li>"
     end
     result.html_safe
   end
 
   def render_pagelet(pagelet, opts = {})
-    if pagelet.onlyif.call opts[:subject]
+    if pagelet.onlyif.call(opts[:subject], self)
       render(pagelet.partial, opts.merge!({ :pagelet => pagelet, controller_name.singularize.to_sym => opts[:subject] })).html_safe
     else
       "".html_safe
