+++ app/views/hosts/puppet/_main_tab_fields.html.erb
@@ -1,0 +1,1 @@
+<%= host_puppet_environment_field form %>
