+++ app/views/hosts/puppet/_puppet_classes_tab.html.erb
@@ -1,0 +1,1 @@
+<%= puppetclasses_tab(@host) %>
