+++ config/initializers/puppet.rb
@@ -1,0 +1,12 @@
+# initialize puppet related pagelets
+mgr = Pagelets::Manager.new "hosts/_form"
+mgr.add_pagelet :main_tabs,
+  :id => :puppet_klasses,
+  :name => _("Puppet Classes"),
+  :partial => "hosts/puppet/puppet_classes_tab",
+  :priority => 100,
+  :onlyif => proc { |host, context| context.instance_eval { accessible_resource(host, :smart_proxy, :name, association: :puppet_proxy).present? } }
+
+mgr.add_pagelet :main_tab_fields,
+  :partial => "hosts/puppet/main_tab_fields",
+  :priority => 100
