+++ test/helpers/pagelets_helper_test.rb
@@ -4,12 +4,13 @@ class PageletsHelperTest < ActionView::TestCase
   include PageletsHelper
 
   setup do
+    @prev_state = Pagelets::Manager.instance_variable_get(:@pagelets).clone
     controller.prepend_view_path File.expand_path('../../static_fixtures/views', __FILE__)
     self.stubs(:virtual_path).returns("nonexisting/path")
   end
 
   teardown do
-    Pagelets::Manager.clear
+    Pagelets::Manager.instance_variable_set :@pagelets, @prev_state
   end
 
   def action_name
